Renamed interfaces to indicate that they are non-functional
