+++ src/drunc/fsm/interfaces/DummyLogbookInterface.py
